+++ types/Webhook/WebhookCreateParameters.d.ts
@@ -1,0 +1,3 @@
+export declare interface IWebhookCreateParameters {
+  url: string;
+}
