+++ types/Webhook/index.d.ts
@@ -1,0 +1,2 @@
+export * from './Webhook';
+export * from './WebhookCreateParameters';
